requirements.txt
weaviate-client==3.9.0
Flask==2.2.2
Pillow
ipykernel
datasets
openai